refactor: Move core functionality into new `Reporter` class
